requirements.txt
tencentcloud-sdk-python==3.0.773

aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-ecs==4.24.26
aliyun-python-sdk-sts==3.1.0

boto3==1.26.29

SQLAlchemy==1.4.44
psycopg2==2.9.5
click==8.1.3
pydantic==1.10.2
requests==2.28.1